backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-dotenv==1.0.0
pandas==2.0.2
numpy==1.24.3
matplotlib==3.7.1
seaborn==0.12.2
transformers==4.29.2
nltk==3.8.1
google-api-python-client==2.86.0
tqdm==4.65.0
httpx==0.24.1